A sample react native app for focusing on various activities
An app with a simple timer functionality which can be used to track time while focusing on different activities. App has all the basic functionality such as Adding activities on which you want to focus, Adding how much you want to focus on the activity, Start and Pausing the timer and reset the timer. App also shows the history of activities which you focused on recently.